workspace: add stopgap test helper to create a colocated workspace
A colocated workspace is one that also happens to have a valid .git file
or directory. Currently, only the default workspace can be colocated, but
we also want secondary workspaces to have this ability.

To create a *secondary* colocated workspace in an existing JJ repo, we
need to add a corresponding git _worktree_.

JJ doesn't have the ability to create git worktrees yet. We want to test
a bunch of code around colocating worktrees with a workspace, assuming
we will build the functionality to actually create one in JJ itself
later. So here's a helper function to do it the really hacky way with
the `git` CLI, moving a .git file into place, and running `git worktree
repair`.